meal_backend/requirements.txt
# FastAPI 관련
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# OpenAI 관련
openai==1.3.3
python-dotenv==1.0.0

# 데이터베이스 및 분석
pandas==2.1.0
numpy==1.25.2
sqlite3-api==0.1

# 벡터 인덱스
llama-index==0.9.16
llama-index-core==0.10.2
llama-index-embeddings-openai==0.1.4

# 기타 유틸리티
python-multipart==0.0.6
requests==2.31.0
